Pushed back the plotting code into weighted_pandas
